Search refactoring
